Made a generic game_base service and exported the session stats calculations into another generic service
